Add files via upload
For my own project it was important that I could upload a firmware from the ESP32 to the Nextion. This because my Nextion SD slot on was fiscal no longer accessible. The ESPNexUpload is the best Nextion upload 1.1 protocol implementation on the web I could found. However, this implementation did not do the job properly. Only when the existing firmware (made with a Nextion V0.58 editor) became corrupt, and I had restarted the display (power off-on). It appears that a standard backup firmware is being activated in the Nextion Display. And that this ‘back-up?’ firmware can properly handle the upload process from the ESPNexUpload library. 
After many wanderings, I finally discovered that a firmware upload from the Nextion editor, with the intervention of a "USB to TTL" converter, to my Nextion display works flawless.  After this discovery I then monitored the serial UART traffic between the two (Editor and Display) with a Logical Analyzer. I saw that that Nextion Editor itself uses undocumented UART commands. Several passages in the Nextion Upload protocol 1.1 also suddenly became more clearer to me.
After the UART data analysis (various test setups made) I reduced what I came across to the essence, and processed it in the ESPNexUpload library. During the development of the changes and extra functionality, I continued to use the Logic Analyzer for checking. This has led, among other things, to improving the existing implementation in several places.